requirements.txt
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.1
scipy==1.2.1
seaborn==0.9.0
plotly==2.0.0
jupyter==1.0.0
fbprophet==0.4.post2
xgboost==0.82
scikit-learn==0.20.3
bokeh==1.0.4
prometheus_client==0.6.0
redis==3.2.1
paho-mqtt==1.4.0